requirements.txt
discord.py==2.5.2
googletrans==4.0.2